Extract GoogleBillingProjectCreator to separate file, update tests
